Add files via upload
Just uploading some of the work I've done so far! I separated the slider into a different widget, and it now displays the value of the slider when the slider position is changed. I'm still working on linking the slider to the "current value" textbox. I also moved the slider and value textbox to the right side of the GUI.